adding react-router and refactoring the app
