+++ src/components/Home.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import MovieCard from './MovieCard'
+
+const Home = (props) => {
+    const {movies} = props
+    return (
+      <div>
+        <MovieCard  movies={movies}/>
+      </div>
+    );
+  }
+
+
+export default Home;
